[fix]: change altitude from int to float

+++ main.py
@@ -195,7 +195,7 @@ def transform_wialon_to_soap(wialon_data):
     if not payload["latitude"]:
         payload["latitude"] = float(last_event["last_event"]["latitude"])
         payload["longitude"] = float(last_event["last_event"]["longitude"])
-        payload["altitude"] = int(last_event["last_event"]["altitude"])
+        payload["altitude"] = float(last_event["last_event"]["altitude"])
         payload["speed"] = int(last_event["last_event"]["speed"])
         payload["odometer"] = int(last_event["last_event"]["odometer"])
         payload["heading"] = int(last_event["last_event"]["heading"])
